+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed error handling for mermaid-cli `^9.1.4`.
+
 ## [2.0.0] - 2022-06-29
 
 ### Breaking Changes
@@ -44,8 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Node v10 support is officially dropped in mermaid-cli 8.10.2**,
   but unofficially, it stills seems to work.
-  
-  Our `dependencies` contain `"@mermaid-js/mermaid-cli":  ^8.9.2"`,
+
+  Our `dependencies` contain `"@mermaid-js/mermaid-cli": ^8.9.2"`,
   however v8.10.2 of mermaid-cli updated to
   [puppeteer v10.0.0](https://github.com/mermaid-js/mermaid-cli/pull/128),
   which officially dropped Node v10 support. However, it still seems to work